move new colors to the colors namespace, start new set of swatches, move old swatches and colors to legacy colors file
