+++ packages/gamut-styles/src/variables.ts
@@ -3,3 +3,6 @@ export * from './variables/responsive';
 export * from './variables/spacing';
 export * from './variables/typography';
 export * from './variables/effects';
+
+// Deprecated variables
+export * from './variables/deprecated-colors';
