+++ packages/gamut-styles/src/variables/typography.ts
@@ -1,4 +1,4 @@
-import { swatches as colorSwatch } from './colors';
+import { legacyColors as colorSwatch } from './deprecated-colors';
 import { pxRem } from '../utilities/pxRem';
 import { base } from './base';
