+++ packages/styleguide/.storybook/Blocks/AboutCell.tsx
@@ -1,5 +1,5 @@
 import { styled } from '@storybook/theming';
-import { colors } from '@codecademy/gamut-styles';
+import { swatches } from '@codecademy/gamut-styles';
 
 // The normal import would be '@storybook/addon-links/react'
 // but it is throwing a typescript error in the current alpha
@@ -25,8 +25,8 @@ export type AboutCellProps = {
 };
 
 const Heading = styled.h2`
-  border-bottom: 1px solid ${colors.gray[200]};
-  color: ${colors.gray[600]};
+  border-bottom: 1px solid ${swatches.gray[200]};
+  color: ${swatches.gray[600]};
   font-size: 1.5rem;
   margin-bottom: 1rem;
   padding-bottom: 0.25rem;
