+++ packages/styleguide/.storybook/Blocks/Box.tsx
@@ -1,10 +1,10 @@
 import styled from '@emotion/styled';
-import { colors } from '@codecademy/gamut-styles';
+import { swatches } from '@codecademy/gamut-styles';
 
 type BoxProps = { size: string };
 
 export const Box = styled.div<BoxProps>`
-  background-color: ${colors.gray[400]};
+  background-color: ${swatches.gray[400]};
   display: inline-block;
   width: ${({ size }) => size};
   height: ${({ size }) => size};
